+++ packages/experience-builder-sdk/src/components/ContentfulContainer.tsx
@@ -1,3 +1,0 @@
-import { ContentfulContainer } from '@contentful/experience-builder-components';
-
-export { ContentfulContainer };