+++ packages/experience-builder-sdk/src/index.ts
@@ -1,6 +1,7 @@
 export { ExperienceRoot } from './ExperienceRoot';
 export { useExperienceBuilder, getValueForBreakpoint, useFetchExperience } from './hooks';
 export { defineComponents } from './core/componentRegistry';
+export { defineDesignTokens } from './core/designTokenRegistry';
 export { calculateNodeDefaultHeight } from './utils/stylesUtils';
 export { checkIfDesignComponent } from './utils/utils';
 export { tryParseMessage, doesMismatchMessageSchema } from './utils/validation';